analyticsAI/requirements.txt
flask==1.0.3
flask_restplus==0.12.1
flask_cors==3.0.7
pyyaml==5.1
requests==2.22.0
nltk==3.4.1
gensim==3.7.3
tika==1.19
flask_basicauth==0.2.0
textblob==0.15.3
scikit-learn==0.21.2
scipy==1.3.0
pandas==0.24.2
google_compute_engine==2.8.13
pymongo==3.8.0
dnspython==1.16.0
spacy==2.1.4
./en_core_web_sm-2.1.0